refactored code to single responsibility
This was a dramatic refactor that converted all logic to be split into single-responsibility functions in order to facilitate testing, as well as to increase maintainability and scalability for the future.

In addition to this, I added type hints, as well as efforted to structure the code as cleanly as possible.